chore(test): move secrets import round-trip test into provider_import_test.go
